Fix RULE-7-0-5 (no-signedness-change-from-promotion) findings
Resolves the majority of RULE-7-0-5 findings in the triage database
using a mix of 3 minimal, unavoidable MISRA deviations, integer-type
widening where a narrow type is purely a local/free-standing counter,
and score::safe_math based code fixes everywhere the narrow width is
load-bearing (bit-packed state, ABI/atomic-width invariants, or
external/template function signatures) -- avoiding an excessive number
of deviation comments for cases that are ordinary runtime signedness
changes and can be made safe in code.

Note: event_data_control_composite.cpp (2 findings, SlotIndexType loop
comparisons) is intentionally left unfixed for now and will be handled
in a follow-up; SlotIndexType is a shared, binding-wide type used to
index EventDataControl::state_slots_ in Shared Memory, so it needs
more careful design discussion before deciding between safe_math and
any alternative.

Deviations (kept to only what truly cannot be expressed in code):
- switch-enum-underlying-type-discriminant: switch/case discriminant
  promotion is a fixed language mechanism and cannot be intercepted by
  any function (3 files, 15 findings).
- ctype-function-argument-promotion: <cctype> functions mandate an int
  parameter per the standard; the promotion is required by the library
  API contract, not by this code (instance_specifier.cpp).
- compile-time-static-assert-limit-check: static_assert-only check, no
  runtime value is ever computed (tracing_runtime.cpp).

Widened to std::uint32_t (rank >= int, so no promotion or signedness
change ever occurs, per RULE-7-0-5's own semantics) instead of using
safe_math, since these are pure local/free-standing counters with no
persisted, shared-memory, or external-ABI role:
- memory_region_map.cpp/.h (both AcquireLatestRegionVersionForRead's
  and AcquireRegionVersionForOverwrite's outer retry-count loops -- each
  is an independent, arbitrary retry counter unrelated to VERSION_COUNT;
  plus VERSION_COUNT itself and AcquireRegionVersionForOverwrite's inner
  loop_idx/VERSION_COUNT loop, once VERSION_COUNT's type was widened from
  std::uint8_t to std::uint32_t. This is safe because VERSION_COUNT is
  only used as a std::array size and in this loop's own bound/modulo --
  its *value* stays 10 and the <= 255 static_assert still holds; the
  actual load-bearing 8-bit invariant lives in latest_known_region_version_
  (an atomic_uint8_t, unchanged) and version_idx, which is still explicitly
  cast back to std::uint8_t before being used/returned. Removing the
  loop_idx/VERSION_COUNT promotion this way also let the loop_idx +
  current_version addition drop safe_math::Add's overflow check entirely:
  with loop_idx in [1, 9] and current_version in [0, 255], the uint32_t
  sum can never overflow)
- transaction_log_local_view.cpp (retry/kRetryCount)
- flag_file_crawler.cpp/.h (max_number_of_retries/current_retry_count)
- tracing_runtime.cpp (impl) + .h (debounce_counter_/kDebounceAfter)
- proxy.cpp (retry_counter/max_retries/kMaxFlockRetries)
- message_passing_service_instance.cpp/.h (kMaxReceiveHandlersPerEvent
  loop and its consumer loop only; the sibling CopyNodeIdentifiers loop
  keeps safe_math since its uint8_t return type is a shared external
  template function signature)

Fixed a latent truncation bug while resolving RULE-7-0-5, rather than
widening around it:
- subscription_subscription_pending_states.cpp (SubscribeEvent): the
  pre-existing code truncated the incoming std::size_t max_sample_count
  down to std::uint8_t *before* comparing it against the stored
  max_sample_count_ (std::uint16_t, legitimately up to 65535). That
  truncation silently corrupted the comparison for any max_sample_count
  above 255, which would have made a legitimate resubscription with the
  same, larger count spuriously fail with kMaxSampleCountNotRealizable.
  Fixed by comparing max_sample_count directly against max_sample_count_
  widened to std::size_t (always lossless, and avoids the promotion
  since std::size_t already has rank >= int), instead of truncating one
  side down to std::uint8_t and then widening the truncated value back
  up to std::uint32_t for the comparison, which fixed the promotion but
  preserved the underlying data-loss bug.

Widened only a local, comparison-only copy (the canonical/stored value
keeps its narrow, wire-format-tied type):
- unix_domain_engine.cpp (ReceiveProtocolMessage): the wire field
  `size` must stay std::uint16_t -- it is read directly into raw
  memory via an iovec (io[1].iov_len = sizeof(size)), so its exact
  2-byte width *is* the protocol written by SendProtocolMessage.
  std::size_t is not a safe substitute either, since its width is
  platform/bitness-dependent and would desync 32-bit/64-bit peers.
  Instead, a local `const std::size_t message_size = size;` copy
  (safe: size's full 0..65535 range is always representable) is used
  for the two same-signedness comparisons that safe_math previously
  guarded. The one genuinely cross-signed comparison left
  (recvmsg's ssize_t byte count vs. the received size) is resolved
  with an explicit, justified static_cast<std::size_t> instead of
  safe_math, since has_value() already guarantees the count is
  non-negative.

Empirically verified via scoped CodeQL databases (per-package, RULE-7-0-5
query only) that none of the widened production sites produce findings.

All other findings -- same-type and cross-type comparisons/arithmetic
alike -- are fixed in code using score::safe_math's CmpEqual/CmpNotEqual/
CmpLess/CmpGreater/CmpLessEqual/CmpGreaterEqual comparison helpers and
Add/Subtract<ReturnMode::kAbortOnError> arithmetic helpers (bounds-safe,
enforced at the call site) across the remaining files, including:
- shared_memory_resource.cpp (uid comparison)
- event_subscription_control.cpp (bit-packed slot/subscriber-count state)
- element_fq_id.cpp, proxy_instance_identifier.cpp,
  skeleton_instance_identifier.cpp, lola_service_instance_identifier.cpp,
  lola_service_instance_id.cpp, unique_method_identifier.cpp
  (equality/ordering operators)
- binding_service_type_deployment_impl.h, lola_event_instance_deployment.cpp,
  trace_point_key.cpp, service_element_tracing_data.h,
  tracing_runtime.cpp (lola bindings), transaction_log_set.cpp,
  sample_allocatee_ptr.h, provider_event_data_control_local_view.cpp,
  subscription_subscribed_states.cpp

Non-integral sub-expressions (enum types, std::optional<T>, map/struct
comparisons) that safe_math cannot express are intentionally left as
plain operators, since no signedness-changing promotion occurs there.

All affected Bazel targets were updated with the necessary
@score_baselibs//score/language/safecpp/safe_math dependency where still
needed (removed where no longer used after widening, e.g.
message_passing_unix_domain), rebuilt, and their existing unit tests
re-run: 126+ regression tests across //score/memory/shared/...,
//score/mw/com/impl/bindings/lola/..., //score/mw/com/impl/tracing/...,
and //score/message_passing/..., 0 regressions.

+++ score/mw/com/impl/bindings/lola/proxy.cpp
@@ -114,13 +114,13 @@ using memory::DataTypeSizeInfo;
 std::unique_ptr<score::memory::shared::FlockMutexAndLock<score::memory::shared::SharedFlockMutex>>
 PlaceSharedLockOnUsageMarkerFileWithRetry(memory::shared::LockFile& service_instance_usage_marker_file,
                                           std::string_view file_path,
-                                          std::uint8_t max_retries)
+                                          std::uint32_t max_retries)
 {
     auto service_instance_usage_mutex_and_lock =
         std::make_unique<score::memory::shared::FlockMutexAndLock<score::memory::shared::SharedFlockMutex>>(
             service_instance_usage_marker_file);
     constexpr std::chrono::milliseconds kRetryBackoffTime{200U};
-    std::uint8_t retry_counter{0U};
+    std::uint32_t retry_counter{0U};
 
     // We use while true and manually break within the loop to prevent sleeping an additional time in case retry_counter
     // exceeds max_retries.
@@ -369,7 +369,7 @@ std::unique_ptr<Proxy> Proxy::Create(const HandleType handle)
         return nullptr;
     }
 
-    constexpr std::uint8_t kMaxFlockRetries{3U};
+    constexpr std::uint32_t kMaxFlockRetries{3U};
     auto service_instance_usage_mutex_and_lock =
         PlaceSharedLockOnUsageMarkerFileWithRetry(service_instance_usage_marker_file.value(),
                                                   std::string_view(service_instance_usage_marker_file_path),
